Dashboard-side changes to start AIC creation chain
refs #5803

Added a CreateAICForm and submit button to the AIP search page that is
populated with the results information.  The results information is posted
to create_aic which creates an entry in a WatchedDirectory to start a
MicroServiceChain.  The directory is populated with files corresponding to
the AIPs that should go in the AIC - the filenames are the UUID, and the
contents are the AIP name.

Also, updated urls.py files to use url() instead of just tuples and did
minor cleanup.