Updated maildir normalization
refs #5955

Added the ability to filter subdirectory on a SIP basis, and used that to
filter by objects/attachments for Maildir in normalization and
identification.

Rather than make many separate MicroServiceChainLinks for Maildir
normalization, added a UnitVariable with the name of the script it applies
to, and a variableValue of {'filterSubDir':'directory to filter'}. The
LinkTaskManager for 'run on every file' checks to see if a UnitVariable for
this script and SIP exists to override filterSubDir, and applies it if one
exists.

The normalization report now only displays files that had a Task run for
them.