Allow transfer metadata entry before components are added
This adds an "Add metadata" button next to the "Start transfer" button,
which allows metadata to be entered while waiting for a component to
be ready. The metadata entered in this form will subsequently be added
to the next component added.

Components can still have their metadata edited after they've been
added.